# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Toke Schools CBT Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teacher/students:
    get:
      operationId: getTeacherStudents
      tags: [staff]
      summary: List all students (requires manage_students permission)
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentRecord"
    post:
      operationId: createTeacherStudent
      tags: [staff]
      summary: Create a new student (requires manage_students permission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
  /teacher/students/{regNumber}:
    put:
      operationId: updateTeacherStudent
      tags: [staff]
      summary: Update a student (requires manage_students permission)
      parameters:
        - name: regNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStudentBody"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentRecord"
    delete:
      operationId: deleteTeacherStudent
      tags: [staff]
      summary: Delete a student (requires manage_students permission)
      parameters:
        - name: regNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teacher/exams:
    get:
      operationId: getTeacherExams
      tags: [staff]
      summary: Get all exams created by this staff member
      responses:
        "200":
          description: List of exams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExamWithStats"
    post:
      operationId: createExam
      tags: [staff]
      summary: Create new exam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "201":
          description: Exam created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
  /teacher/exams/{examId}/toggle-results:
    patch:
      operationId: toggleExamResults
      tags: [staff]
      summary: Enable or disable result visibility for an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleExamResultsBody"
      responses:
        "200":
          description: Results visibility updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teacher/exams/{examId}:
    get:
      operationId: getTeacherExam
      tags: [staff]
      summary: Get exam details
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamWithQuestions"
    put:
      operationId: updateExam
      tags: [staff]
      summary: Update exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExamBody"
      responses:
        "200":
          description: Exam updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Exam"
    delete:
      operationId: deleteExam
      tags: [staff]
      summary: Delete exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Exam deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teacher/exams/{examId}/questions:
    get:
      operationId: getExamQuestions
      tags: [questions]
      summary: Get questions for an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of questions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"
    post:
      operationId: createQuestion
      tags: [questions]
      summary: Add a question to an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "201":
          description: Question created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
  /teacher/exams/{examId}/questions/{questionId}:
    put:
      operationId: updateQuestion
      tags: [questions]
      summary: Update a question
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuestionBody"
      responses:
        "200":
          description: Question updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Question"
    delete:
      operationId: deleteQuestion
      tags: [questions]
      summary: Delete a question
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Question deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teacher/exams/{examId}/results/{resultId}:
    delete:
      operationId: deleteExamResult
      tags: [staff]
      summary: Remove a student's result from an exam, making it available to retake
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
        - name: resultId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Result removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teacher/exams/{examId}/results:
    get:
      operationId: getExamResults
      tags: [results]
      summary: Get all student results for an exam
      parameters:
        - name: examId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultWithStudent"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
...
    StudentRecord:
      type: object
      properties:
        regNumber:
          type: string
        name:
          type: string
        class:
          type: string
        isDefaultPassword:
          type: boolean
        studentRole:
          type: string
          nullable: true
        profilePicture:
          type: string
          nullable: true
      required: [regNumber, name, class, isDefaultPassword]
    CreateStudentBody:
      type: object
      properties:
        regNumber:
          type: string
        name:
          type: string
        class:
          type: string
        password:
          type: string
        studentRole:
          type: string
          nullable: true
      required: [regNumber, name, class, password]
    UpdateStudentBody:
      type: object
      properties:
        name:
          type: string
        class:
          type: string
        password:
          type: string
          nullable: true
        resetPassword:
          type: boolean
          nullable: true
        studentRole:
          type: string
          nullable: true
      required: [name, class]
...
    Exam:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        class:
          type: string
        durationMinutes:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        createdBy:
          type: string
        questionCount:
          type: integer
        alreadySubmitted:
          type: boolean
          nullable: true
        resultsEnabled:
          type: boolean
          nullable: true
      required: [id, subject, class, durationMinutes, createdBy, questionCount]
    ExamWithStats:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        class:
          type: string
        durationMinutes:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        createdBy:
          type: string
        questionCount:
          type: integer
        attemptCount:
          type: integer
        averageScore:
          type: number
          nullable: true
        resultsEnabled:
          type: boolean
      required: [id, subject, class, durationMinutes, createdBy, questionCount, attemptCount, resultsEnabled]
    Question:
      type: object
      properties:
        id:
          type: integer
        examId:
          type: integer
        questionText:
          type: string
        optionA:
          type: string
        optionB:
          type: string
        optionC:
          type: string
        optionD:
          type: string
        correctOption:
          type: string
          enum: [A, B, C, D]
      required: [id, examId, questionText, optionA, optionB, optionC, optionD, correctOption]
    ExamWithQuestions:
      type: object
      properties:
        id:
          type: integer
        subject:
          type: string
        class:
          type: string
        durationMinutes:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
        createdBy:
          type: string
        resultsEnabled:
          type: boolean
          nullable: true
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
      required: [id, subject, class, durationMinutes, createdBy, questions]
    CreateExamBody:
      type: object
      properties:
        subject:
          type: string
        class:
          type: string
        durationMinutes:
          type: integer
        startTime:
          type: string
          nullable: true
        endTime:
          type: string
          nullable: true
      required: [subject, class, durationMinutes]
...
    ToggleExamResultsBody:
      type: object
      properties:
        enabled:
          type: boolean
      required: [enabled]